PDEV-23 change FaultCode to use shuffleboard, remove duplicate entries
